refactor: journal-photos.entity primary key 추가
